[form] Rename `FormContent`, export `SubmitButton`
